+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
-## 1.0.3 - 2023-08-21
-* Bucket and Object Information
+## 1.0.4 - 2023-08-22
+* s3_security.go Created
+* Bucket Policies and Permissions